Replace comparison of enums of different types with comparison of ints.
git-svn-id: svn+ssh://source-control.beyondtrust.com/sources/repo/bt-dev/trunk/Platform/src/linux@44372 9fc1c11e-11b2-4e14-94d8-6c5e05775969

+++ lsass/logging_r/logging.c
@@ -57,7 +57,7 @@ LsaLwLogMessage(
     )
 {
     LSA_LOCK_LOGGER;
-    if (gpfnLogger && (gLsaMaxLogLevel >= level))
+    if (gpfnLogger && ((int)gLsaMaxLogLevel >= (int)level))
     {
         _LSA_LOG_WITH_THREAD(level, "%s", pszMessage);
     }
@@ -73,7 +73,7 @@ LsaLwpsLogMessage(
     )
 {
     LSA_LOCK_LOGGER;
-    if (gpfnLogger && (gLsaMaxLogLevel >= level))
+    if (gpfnLogger && ((int)gLsaMaxLogLevel >= (int)level))
     {
         _LSA_LOG_WITH_THREAD(level, "%s", pszMessage);
     }
